Updated the base classes files and modified the install and uninstall scripts.

+++ install.py
@@ -486,7 +486,8 @@ def cleanUp():
                         os.remove(apiname)
                 for apiname in glob.glob(
                         os.path.join(apidir, progLanguage.lower(), "*.bas")):
-                    os.remove(apiname)
+                    if os.path.basename(apiname) != "eric6.bas":
+                        os.remove(apiname)
         except AttributeError:
             pass
